:fire: removing mixins implementation :recycle: update based on mynavi integration testing :recycle: changing to handle single record
