+++ igata/handlers/__init__.py
@@ -3,7 +3,6 @@ from collections import defaultdict
 from .aws.input.s3 import S3BucketImageInputCtxManager
 from .aws.input.sqs import SQSMessageS3InputCSVPandasDataFrameCtxManager, SQSMessageS3InputImageCtxManager
 from .aws.output.dynamodb import DynamodbOutputCtxManager
-from .aws.output.mixins.dyanamodb import DynamodbRequestUpdateMixIn
 from .aws.output.s3 import S3BucketPandasDataFrameCsvFileOutputCtxManager
 from .aws.output.sqs import SQSRecordOutputCtxManager
 
@@ -36,5 +35,3 @@ for name, ctxmgr_class in OUTPUT_CONTEXT_MANAGERS.items():
     for required_arguement_name in ctxmgr_class.required_kwargs():
         required_envar_name = f"{output_ctxmgr_envar_prefix}{required_arguement_name.upper()}"
         OUTPUT_CONTEXT_MANAGER_REQUIRED_ENVARS[name].append(required_envar_name)
-
-OUTPUT_CONTEXT_MANAGER_MIXINS = {"DynamodbRequestUpdateMixIn": DynamodbRequestUpdateMixIn}
