--- tests/dummypredictor/mixins.py
@@ -1,6 +1,0 @@
-from igata.handlers.aws.output.mixins import PostPredictHookMixInBase
-
-
-class DummyMixin(PostPredictHookMixInBase):
-    def mixin_method(self) -> bool:
-        return True
