--- tests/handlers/output/test_handlers_aws_output_mixins_dynamodbrequestupdatepostpredictmixin.py
@@ -1,3 +1,0 @@
-# TODO: Finish!
-# def test_dynamodbrequestupdatepostpredictmixin():
-#    raise NotImplementedError
